Add: `useCurrentToken` specialized hook
