+++ src/hooks/index.ts
@@ -49,6 +49,7 @@ export { default as useExtensionData } from './useExtensionData';
 export { default as useExtensionsData } from './useExtensionsData';
 export * from './useCanInteractWithColony';
 export { default as useColonyFundsClaims } from './useColonyFundsClaims';
+export { default as useCurrentSelectedToken } from './useCurrentSelectedToken';
 
 /* Used in cases where we need to memoize the transformed output of any data.
  * Transform function has to be pure, obviously
